requirements.txt
paho-mqtt==1.4.0
colorama==0.3.9
sdnotify==0.3.1
pymodbus==2.3.0
Unidecode==0.4.21
